+++ src/senaite/core/adapters/localroles.py
@@ -20,28 +20,32 @@
 
 import six
 from bika.lims import api
+from bika.lims import APIError
 from bika.lims import logger
 from bika.lims.utils import get_client
 from borg.localrole.default_adapter import DefaultLocalRoleAdapter
-from plone.memoize import ram
+from plone.memoize.request import cache
 from senaite.core.behaviors import IClientShareableBehavior
 from senaite.core.interfaces import IDynamicLocalRoles
 from zope.component import getAdapters
 from zope.interface import implementer
 
 
-def _getRolesInContext_cachekey(method, self, context, principal_id):
-    """Function that generates the key for volatile caching
+def _request_cache(method, *args):
+    """Decorator that ensures the call to the given method with the given
+    arguments only takes place once within same request
     """
-    # We need the cachekey to change when global roles of a given user change
-    user = api.get_user(principal_id)
-    roles = user and ":".join(sorted(user.getRoles())) or ""
-    return ".".join([
-        principal_id,
-        roles,
-        api.get_path(context),
-        api.get_modification_date(context).ISO(),
-    ])
+    # Extract the path of the context from the instance
+    try:
+        path = api.get_path(args[0].context)
+    except APIError:
+        path = "no-context"
+
+    return [
+        method.__name__,
+        path,
+        args[1:],
+    ]
 
 
 class DynamicLocalRoleAdapter(DefaultLocalRoleAdapter):
@@ -50,7 +54,11 @@ class DynamicLocalRoleAdapter(DefaultLocalRoleAdapter):
     current traverse path
     """
 
-    @ram.cache(_getRolesInContext_cachekey)
+    @property
+    def request(self):
+        # Fixture for tests that do not have a regular request!!!
+        return api.get_request() or api.get_test_request()
+
     def getRolesInContext(self, context, principal_id):
         """Returns the dynamically calculated 'local' roles for the given
         principal and context
@@ -58,10 +66,6 @@ class DynamicLocalRoleAdapter(DefaultLocalRoleAdapter):
         @param principal_id: User login id
         @return List of dynamically calculated local-roles for user and context
         """
-        if not api.get_user(principal_id):
-            # principal_id can be a group name, but we consider users only
-            return []
-
         roles = set()
         path = api.get_path(context)
         adapters = getAdapters((context,), IDynamicLocalRoles)
@@ -73,6 +77,7 @@ class DynamicLocalRoleAdapter(DefaultLocalRoleAdapter):
             roles.update(local_roles)
         return list(roles)
 
+    @cache(get_key=_request_cache, get_request='self.request')
     def getRoles(self, principal_id):
         """Returns both non-local and local roles for the given principal in
         current context
@@ -84,6 +89,10 @@ class DynamicLocalRoleAdapter(DefaultLocalRoleAdapter):
             # We only apply dynamic local roles to valid objects
             return default_roles[:]
 
+        if principal_id not in self.getMemberIds():
+            # We only apply dynamic local roles to existing users
+            return default_roles[:]
+
         # Extend with dynamically computed roles
         dynamic_roles = self.getRolesInContext(self.context, principal_id)
         return list(set(default_roles + dynamic_roles))
@@ -92,12 +101,18 @@ class DynamicLocalRoleAdapter(DefaultLocalRoleAdapter):
         roles = {}
         # Iterate through all members to extract their dynamic local role for
         # current context
-        mtool = api.get_tool("portal_membership")
-        for principal_id in mtool.listMemberIds():
+        for principal_id in self.getMemberIds():
             user_roles = self.getRoles(principal_id)
             if user_roles:
                 roles.update({principal_id: user_roles})
         return six.iteritems(roles)
+
+    @cache(get_key=_request_cache, get_request='self.request')
+    def getMemberIds(self):
+        """Return the list of user ids
+        """
+        mtool = api.get_tool("portal_membership")
+        return mtool.listMemberIds()
 
 
 @implementer(IDynamicLocalRoles)